Add files via upload
In this project, I applied the skills I learned in Introduction to Python and Intermediate Python to solve a real-world data science problem. We'll discover if Netflix’s movies are getting shorter over time and which guest stars appear in the most popular episode of "The Office", using everything from lists and loops to pandas and matplotlib.

I also gained experience in an essential data science skill — exploratory data analysis. This allows me to perform critical tasks such as manipulating raw data and drawing conclusions from plots created of the data.